% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_decomposition.R
\name{print_decomposition}
\alias{print_decomposition}
\title{Print the decomposition}
\usage{
print_decomposition(x, format = "latex", digits = 3,
  decimal.mark = getOption("OutDec"), booktabs = TRUE, ...)
}
\arguments{
\item{x}{the object to print.}

\item{format}{output format: \code{"latex"} or \code{"html"}.}

\item{digits}{number of digits after the decimal point.}

\item{decimal.mark}{the character to be used to indicate the numeric decimal point.}

\item{booktabs}{boolean indicating whether to use or not the booktabs package (when \code{format = "latex"}).}

\item{...}{unused.}
}
\description{
Print the decomposition
}
